Moved Exporter Set into its own file
