+++ ZAPD/Globals.h
@@ -5,6 +5,7 @@
 #include <vector>
 #include "GameConfig.h"
 #include "ZFile.h"
+#include "ExporterSet.h"
 
 class ZRoom;
 
@@ -13,29 +14,6 @@ enum class VerbosityLevel
 	VERBOSITY_SILENT,
 	VERBOSITY_INFO,
 	VERBOSITY_DEBUG
-};
-
-typedef void (*ExporterSetFunc)(ZFile*);
-typedef bool (*ExporterSetFuncBool)(ZFileMode fileMode);
-typedef void (*ExporterSetFuncVoid)(int argc, char* argv[], int& i);
-typedef void (*ExporterSetFuncVoid2)(const std::string& buildMode, ZFileMode& fileMode);
-typedef void (*ExporterSetFuncVoid3)();
-typedef void (*ExporterSetResSave)(ZResource* res, BinaryWriter& writer);
-
-class ExporterSet
-{
-public:
-	~ExporterSet();
-
-	std::map<ZResourceType, ZResourceExporter*> exporters;
-	ExporterSetFuncVoid parseArgsFunc = nullptr;
-	ExporterSetFuncVoid2 parseFileModeFunc = nullptr;
-	ExporterSetFuncBool processFileModeFunc = nullptr;
-	ExporterSetFunc beginFileFunc = nullptr;
-	ExporterSetFunc endFileFunc = nullptr;
-	ExporterSetFuncVoid3 beginXMLFunc = nullptr;
-	ExporterSetFuncVoid3 endXMLFunc = nullptr;
-	ExporterSetResSave resSaveFunc = nullptr;
 };
 
 class Globals
