+++ ZAPD/ZAPD.vcxproj
@@ -273,6 +273,7 @@ mkdir build\ZAPD
     <ClInclude Include="CrashHandler.h" />
     <ClInclude Include="CRC32.h" />
     <ClInclude Include="Declaration.h" />
+    <ClInclude Include="ExporterSet.h" />
     <ClInclude Include="GameConfig.h" />
     <ClInclude Include="Globals.h" />
     <ClInclude Include="ImageBackend.h" />
