+++ ZAPD/ZAPD.vcxproj.filters
@@ -1,4 +1,4 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
+<?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <ItemGroup>
     <Filter Include="Source Files">
@@ -561,7 +561,10 @@
       <Filter>Header Files\Z64</Filter>
     </ClInclude>
     <ClInclude Include="ZActorList.h">
-      <Filter>Header Files\Z64</Filter>
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="ExporterSet.h">
+      <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="ZWaterbox.h">
       <Filter>Header Files</Filter>
